Reorganized odometry code. Experimental vision odometry added
