+++ src/main/java/frc/robot/RobotContainer.java
@@ -37,8 +37,9 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 public class RobotContainer {
   // The robot's subsystems and commands are defined here...
   private final ExampleSubsystem m_exampleSubsystem = new ExampleSubsystem();
+  private final Odometry m_Odometry =new Odometry();
 
-  private final Drivetrain       m_drivetrain =       new Drivetrain();
+  private final Drivetrain       m_drivetrain =       new Drivetrain(m_Odometry);
   private final Intake           m_intake =           new Intake();
   private final Staging          m_staging =          new Staging();
   private final TurretRotator    m_turretRotator =    new TurretRotator();
