+++ src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -48,11 +48,15 @@ public class Drivetrain extends SubsystemBase {
   private double v;
   private double lastPosition;
   private double currentAngle;
+
+  
+
   private Pose2d robotPosition;
   private DifferentialDriveOdometry robotOdometry = new DifferentialDriveOdometry(Rotation2d.fromDegrees(0.0));
   private Rotation2d robotRotation;
   private Translation2d robotTranslation;
   Drivetrain m_drivetrain;
+  Odometry m_odometry;
 
 
   private double lP, lI, lD, lF, rP, rI, rD, rF, lSP, rSP;
@@ -65,7 +69,8 @@ public class Drivetrain extends SubsystemBase {
   /**
    * Creates a new Drivetrain.
    */
-  public Drivetrain() {
+  public Drivetrain(Odometry odometry) {
+    m_odometry = odometry;
     leftLeader.setInverted(false);
     rightLeader.setInverted(true);
 
@@ -142,24 +147,7 @@ public class Drivetrain extends SubsystemBase {
       }
     } 
     TankDrive(leftMotorOutput, rightMotorOutput);
-    // v = GetSpeedInMetersPerCentisecond(lastPosition);
-    // lastPosition = GetEncodersTotal();
-
-    // currentAngle = Variables.getInstance().getGyroAngle();
-    // robotRotation = Rotation2d.fromDegrees(currentAngle);
-
-    // //worldVector = worldVector.se
-    // LeftDistanceMeters = GetLeftDistanceMeters();
-    // RightDistanceMeters = GetRightDistanceMeters();
-    // robotOdometry.update(robotRotation, GetLeftDistanceMeters(), GetRightDistanceMeters());
-    // robotPosition =  robotOdometry.getPoseMeters();
-    // robotTranslation = robotPosition.getTranslation();
-    // //if (slow_print > 1000) {
-    //   System.out.printf("X:%8.3f  Y:%8.3f  Angle:%8.3f  Left:%8f  Right:%8f\n", robotTranslation.getX(),robotTranslation.getY(), currentAngle, LeftDistanceMeters, RightDistanceMeters);
-    //   slow_print = 0;
-    //} 
-    //++slow_print;
-    //if (slow_print < 0) slow_print = 0;
+    
   }
 
   public void TankDrive(double leftValue, double rightValue){
@@ -171,25 +159,27 @@ public class Drivetrain extends SubsystemBase {
     TankDrive(0, 0);
   }
 
-  public void PrintLocation(){
-    // v = GetSpeedInMetersPerCentisecond(lastPosition);
-    // lastPosition = GetEncodersTotal();
-
-    // currentAngle = Variables.getInstance().getGyroAngle();
-    // robotRotation = Rotation2d.fromDegrees(currentAngle);
-
-    // //worldVector = worldVector.se
-    // robotOdometry.update(robotRotation, GetLeftDistanceMeters(), GetRightDistanceMeters());
-    // robotPosition =  robotOdometry.getPoseMeters();
-    // robotTranslation = robotPosition.getTranslation();
-    // System.out.println("x Translation" + robotTranslation.getX());
-    // System.out.println("y Translation" + robotTranslation.getY());
-  }
+  
 
   @Override
   public void periodic() {
     //System.out.println(leftController.getIAccum());
     // This method will be called once per scheduler run
+    if(m_odometry.tv.isValid() && !m_odometry.wasValid){
+      m_odometry.updateFromReflectors();
+      m_odometry.wasValid = true;
+    }
+    robotOdometry.update(robotRotation, m_odometry.resetableEncoderLeft, m_odometry.resetableEncoderRight);
+    currentAngle = Variables.getInstance().getGyroAngle();
+    robotRotation = Rotation2d.fromDegrees(currentAngle);
+
+    System.out.println("x Translation" + robotTranslation.getX());
+    System.out.println("y Translation" + robotTranslation.getY());
+
+    m_odometry.lastDistanceLeft = m_drivetrain.GetLeftEncoder();
+    m_odometry.lastDistanceRight = m_drivetrain.getRightEncoder();
+
+    lastPosition = GetEncodersTotal();
   }
   public double getRightEncoder() {
     return rightEncoder.getPosition();
@@ -206,4 +196,7 @@ public class Drivetrain extends SubsystemBase {
   public void resetRightEncoder(){
     rightEncoder.setPosition(0);
   }
+
+  
 }
+
